Add 'Main.WebHome' with redirection and panel
The 'Main.WebHome' will display the list of subwikis (personal or collaborative
spaces) available to the current user.  If the user doesn't exists in XWiki, it
creates it and add it to the needed groups.  Finally, when going to a subwiki
through the 'EESCCode.CSRedirect' redirection page, the subwiki is created if it
doesn't exists, and if it exists, the user is redirect on the main page of this
wiki.